Move new maps to v2 route
